Fix impoper use of next inside of user defined function
According to GAWK documentation:

According to the POSIX standard, the behavior is undefined if the next
statement is used in a BEGIN or END rule. gawk treats it as a syntax
error. Although POSIX does not disallow it, most other awk
implementations don’t allow the next statement inside function bodies
(see User-Defined Functions). Just as with any other next statement, a
next statement inside a function body reads the next record and starts
processing it with the first rule in the program.